+++ src/torchjd/autojac/mtl_backward.py
@@ -154,6 +154,10 @@ def mtl_backward(
         ``mtl_backward`` relies on a usage of ``torch.vmap`` that is not compatible with compiled
         functions. The arguments of ``mtl_backward`` should thus not come from a compiled model.
         Check https://github.com/pytorch/pytorch/issues/138422 for the status of this issue.
+
+    .. warning::
+        Because of a limitation of ``torch.vmap``, tensors in the computation graph of the
+        ``features`` parameter should not have their ``retains_grad`` parameter set to ``True``.
     """
     if shared_params is None:
         shared_params, _ = _get_tasks_shared_params(losses)